+++ CHANGELOG.md
@@ -116,6 +116,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Application\Model\OrderArticle::$_aOrderCache`
 - `\OxidEsales\EshopCommunity\Core\Module\ModuleTemplatePathCalculator`
 - `\OxidEsales\EshopCommunity\Core\Config::getModuleTemplatePathCalculator()`
+- `\OxidEsales\EshopCommunity\Core\Module\ModuleList::getModuleTemplates()`
+- `\OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler\TemplatesModuleSettingHandler`
+- `\OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject\ShopConfigurationSetting::MODULE_TEMPLATES`
 - Removed deprecated getSession and setSession usages around the code
 - `OxidEsales\Eshop\Application\Model\UserPayment::$_sPaymentKey`
 - `OxidEsales\Eshop\Application\Model\UserPayment::getPaymentKey()`